Tests in the source directories are now discovered.
Previously, tests that were placed in the --source directories were
not discovered correctly. Instead, the script was just scanning for all
tests in the current directory. You needed to use the -p option to force
the script to look in the right directories.

This will now correctly use the --source directories to look for tests.
The -p option may still be used to filter tests, however they must be
located in the --source directories or a non-ignored subdirectory.